pyproject.toml
# Copyright 2021 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "opentelemetry-collector-k8s"
version = "0.0"
requires-python = "~=3.8"

dependencies = [
  "ops>=2.17",
  "pyyaml",
  "urllib3",
  "jsonschema",
  "cryptography",
  "lightkube>=0.11",
  "cosl",
]

[project.optional-dependencies]
dev = [
  # Linting
  "ruff",
  "codespell",
  # Static
  "pyright",
  # Unit
  "pytest",
  "coverage[toml]",
  # Integration
  "juju<=3.3.0,>=3.0",
  "websockets<14.0",
  "pytest-operator>=0.39",
  "ops[testing]",
  "sh",
  "tenacity",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

# Linting tools configuration
[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]

[tool.ruff.lint]
select = ["E", "W", "F", "C", "N", "R", "D"]
# Ignore E501 because using black creates errors with this
# Ignore D107 Missing docstring in __init__
ignore = ["W505", "E501", "D107", "C901", "N818", "RET504"]
# D100, D101, D102, D103: Ignore missing docstrings in tests
per-file-ignores = {"tests/*" = ["D100","D101","D102","D103"]}

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
extraPaths = ["lib"]
pythonVersion = "3.8"
pythonPlatform = "All"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--tb=native --verbose --capture=no --log-cli-level=INFO"

[tool.codespell]
skip = ".git,.tox,build,venv*"
ignore-words-list = "assertIn"
